+++ spot/plugins/CPanel.py
@@ -130,7 +130,7 @@ class CPanel(GingaPlugin.GlobalPlugin):
         scr.set_widget(self.w.stk)
         top.add_widget(scr, stretch=1)
 
-        top.add_widget(Widgets.Label(''), stretch=1)
+        #top.add_widget(Widgets.Label(''), stretch=1)
 
         btns = Widgets.HBox()
         btns.set_border_width(4)